requirements.txt
Django==5.0
django-environ==0.11.2
django-nextjs==2.4.0
psycopg2==2.9.9
